Amend PNG image fallback code
The "onerror" handler for the SVG image that aims at replacing the image with a
PNG version created an infinite loop in the (unlikely, granted) event when the
client lost her network connection after having loaded the spec but before
having loaded the SVG image.

In such a scenario, the fetch on the image would trigger the handler, which
would change the image source to the URL of the PNG image, which would trigger
another fetch, which would trigger the handler, which would change the source
to the same value but that's enough to trigger another fetch, etc.

This fix only changes the source URL once.